message api correction

+++ src/main/java/com/chatop/dto/MessagesDto.java
@@ -3,13 +3,13 @@ package com.chatop.dto;
 
 public class MessagesDto {
     private String message;
-    private Long rental_id;
-    private Long user_id;
+    private int rental_id;
+    private int user_id;
 
     public MessagesDto() {
     }
     
-    public MessagesDto(String message, Long user_id, Long rental_id) {
+    public MessagesDto(String message, int user_id, int rental_id) {
         this.message = message;
         this.rental_id = rental_id;
         this.user_id = user_id;
@@ -21,16 +21,16 @@ public class MessagesDto {
     public void setMessage(String message) {
         this.message = message;
     }
-    public Long getUser_id() {
+    public int getUser_id() {
         return user_id;
     }
-    public void setUser_id(Long user_id) {
+    public void setUser_id(int user_id) {
         this.user_id = user_id;
     }
-    public Long getRental_id() {
+    public int getRental_id() {
         return rental_id;
     }
-    public void setRental_id(Long rental_id) {
+    public void setRental_id(int rental_id) {
         this.rental_id = rental_id;
     }
